Adjust folders
