+++ Assets/Scenes/StartScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c74e337e4d1c34e3a872633d613252d6
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
