+++ Assets/Scripts/Core/Services.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7cf021e8e113147138c6be41e39d38d8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
