+++ Assets/Scripts/Core/Services/IAudioService.cs
@@ -1,0 +1,8 @@
+namespace Core.Services
+{
+    public interface IAudioService
+    {
+        bool AudioEnabled { get; }
+        void SetAudioEnabled(object anchor, bool isEnabled);
+    }
+}
