+++ Assets/Scripts/DungeonGeneration/DungeonCrawler.cs
@@ -1,19 +1,21 @@
-﻿using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using UnityEngine;
 
-public class DungeonCrawler
+namespace DungeonGeneration
 {
-    public Vector2Int Position { get; set; }
-    public DungeonCrawler(Vector2Int startPos)
+    public class DungeonCrawler
     {
-        Position = startPos;
-    }
+        public Vector2Int Position { get; set; }
+        public DungeonCrawler(Vector2Int startPos)
+        {
+            Position = startPos;
+        }
 
-    public Vector2Int Move(Dictionary<Direction, Vector2Int> directionMovementMap)
-    {
-        Direction toMove = (Direction)Random.Range(0, directionMovementMap.Count);
-        Position += directionMovementMap[toMove];
-        return Position;
+        public Vector2Int Move(Dictionary<Direction, Vector2Int> directionMovementMap)
+        {
+            Direction toMove = (Direction)Random.Range(0, directionMovementMap.Count);
+            Position += directionMovementMap[toMove];
+            return Position;
+        }
     }
 }
