+++ Assets/Scripts/DungeonGeneration/DungeonGenerationData.cs
@@ -1,9 +1,12 @@
 ﻿using UnityEngine;
 
-[CreateAssetMenu(fileName = "DungeonGenerationData.asset", menuName = "DungeonGenerationData/Dungeon Data")]
-public class DungeonGenerationData : ScriptableObject
+namespace DungeonGeneration
 {
-    public int numberOfCrawlers;
-    public int iterationMin;
-    public int iterationMax;
+    [CreateAssetMenu(fileName = "DungeonGenerationData.asset", menuName = "DungeonGenerationData/Dungeon Data")]
+    public class DungeonGenerationData : ScriptableObject
+    {
+        public int numberOfCrawlers;
+        public int iterationMin;
+        public int iterationMax;
+    }
 }
