+++ Assets/Scripts/DungeonGeneration/DungeonGenerator.cs
@@ -1,25 +1,27 @@
-﻿using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using UnityEngine;
 
-public class DungeonGenerator : MonoBehaviour
+namespace DungeonGeneration
 {
-    public DungeonGenerationData dungeonGenerationData;
-    private List<Vector2Int> dungeonRooms;
-
-    private void Start()
+    public class DungeonGenerator : MonoBehaviour
     {
-        dungeonRooms = DungeonCrawlerController.GenerateDungeon(dungeonGenerationData);
-        SpawnRooms(dungeonRooms);
-    }
+        public DungeonGenerationData dungeonGenerationData;
+        private List<Vector2Int> dungeonRooms;
 
-    private void SpawnRooms(IEnumerable<Vector2Int> rooms)
-    {
-        RoomController.instance.LoadRoom("Start", 0, 0);
-        ;
-        foreach(Vector2Int roomLocation in rooms)
+        private void Start()
         {
-            RoomController.instance.LoadRoom(RoomController.instance.GetRandomRoomName(), roomLocation.x, roomLocation.y);
+            dungeonRooms = DungeonCrawlerController.GenerateDungeon(dungeonGenerationData);
+            SpawnRooms(dungeonRooms);
+        }
+
+        private void SpawnRooms(IEnumerable<Vector2Int> rooms)
+        {
+            RoomController.instance.LoadRoom("Start", 0, 0);
+            ;
+            foreach(Vector2Int roomLocation in rooms)
+            {
+                RoomController.instance.LoadRoom(RoomController.instance.GetRandomRoomName(), roomLocation.x, roomLocation.y);
+            }
         }
     }
 }
