+++ Assets/Scripts/DungeonGeneration/Room.cs
@@ -1,149 +1,151 @@
-﻿using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using UnityEngine;
 
-public class Room : MonoBehaviour
+namespace DungeonGeneration
 {
-
-    public int Width;
-    public int Height;
-    public int X;
-    public int Y;
-
-    private bool updatedDoors = false;
-
-    public Room(int x, int y)
+    public class Room : MonoBehaviour
     {
-        X = x;
-        Y = y;
-    }
 
-    private Door leftDoor;
-    private Door rightDoor;
-    private Door topDoor;
-    private Door bottomDoor;
+        public int Width;
+        public int Height;
+        public int X;
+        public int Y;
 
-    public List<Door> doors = new List<Door>();
+        private bool updatedDoors = false;
 
-    // Start is called before the first frame update
-    void Start()
-    {
-        if(RoomController.instance == null)
+        public Room(int x, int y)
         {
-            Debug.Log("You pressed play in the wrong scene!");
-            return;
+            X = x;
+            Y = y;
         }
 
-        Door[] ds = GetComponentsInChildren<Door>(); 
-        foreach(Door d in ds)
+        private Door leftDoor;
+        private Door rightDoor;
+        private Door topDoor;
+        private Door bottomDoor;
+
+        public List<Door> doors = new List<Door>();
+
+        // Start is called before the first frame update
+        void Start()
         {
-            doors.Add(d);
-            switch(d.doorType)
+            if(RoomController.instance == null)
             {
-                case Door.DoorType.right:
-                rightDoor = d;
-                break;
-                case Door.DoorType.left:
-                leftDoor = d;
-                break;
-                case Door.DoorType.top:
-                topDoor = d;
-                break;
-                case Door.DoorType.bottom:
-                bottomDoor = d;
-                break;
+                Debug.Log("You pressed play in the wrong scene!");
+                return;
+            }
+
+            Door[] ds = GetComponentsInChildren<Door>(); 
+            foreach(Door d in ds)
+            {
+                doors.Add(d);
+                switch(d.doorType)
+                {
+                    case Door.DoorType.right:
+                        rightDoor = d;
+                        break;
+                    case Door.DoorType.left:
+                        leftDoor = d;
+                        break;
+                    case Door.DoorType.top:
+                        topDoor = d;
+                        break;
+                    case Door.DoorType.bottom:
+                        bottomDoor = d;
+                        break;
+                }
+            }
+
+            RoomController.instance.RegisterRoom(this);
+        }
+
+        void Update()
+        {
+            if(name.Contains("End") && !updatedDoors)
+            {
+                RemoveUnconnectedDoors();
+                updatedDoors = true;
             }
         }
 
-        RoomController.instance.RegisterRoom(this);
-    }
-
-    void Update()
-    {
-        if(name.Contains("End") && !updatedDoors)
+        public void RemoveUnconnectedDoors()
         {
-            RemoveUnconnectedDoors();
-            updatedDoors = true;
-        }
-    }
-
-    public void RemoveUnconnectedDoors()
-    {
-        Debug.Log("removing doors");
-        foreach(Door door in doors)
-        {
-            switch(door.doorType)
+            Debug.Log("removing doors");
+            foreach(Door door in doors)
             {
-                case Door.DoorType.right:
-                    if(GetRight() == null)
-                        door.gameObject.SetActive(false);
-                    break;
-                case Door.DoorType.left:
-                    if(GetLeft() == null)
-                        door.gameObject.SetActive(false);
-                    break;
-                case Door.DoorType.top:
-                    if(GetTop() == null)
-                        door.gameObject.SetActive(false);
-                    break;
-                case Door.DoorType.bottom:
-                if(GetBottom() == null)
-                        door.gameObject.SetActive(false);
-                break;
+                switch(door.doorType)
+                {
+                    case Door.DoorType.right:
+                        if(GetRight() == null)
+                            door.gameObject.SetActive(false);
+                        break;
+                    case Door.DoorType.left:
+                        if(GetLeft() == null)
+                            door.gameObject.SetActive(false);
+                        break;
+                    case Door.DoorType.top:
+                        if(GetTop() == null)
+                            door.gameObject.SetActive(false);
+                        break;
+                    case Door.DoorType.bottom:
+                        if(GetBottom() == null)
+                            door.gameObject.SetActive(false);
+                        break;
+                }
             }
         }
-    }
 
-    public Room GetRight()
-    {
-        if(RoomController.instance.DoesRoomExist(X + 1, Y))
+        public Room GetRight()
         {
-            return RoomController.instance.FindRoom(X + 1, Y);
+            if(RoomController.instance.DoesRoomExist(X + 1, Y))
+            {
+                return RoomController.instance.FindRoom(X + 1, Y);
+            }
+            return null;
         }
-        return null;
-    }
-    public Room GetLeft()
-    {
-        if(RoomController.instance.DoesRoomExist(X - 1, Y))
+        public Room GetLeft()
         {
-            return RoomController.instance.FindRoom(X - 1, Y);
+            if(RoomController.instance.DoesRoomExist(X - 1, Y))
+            {
+                return RoomController.instance.FindRoom(X - 1, Y);
+            }
+            return null;
         }
-        return null;
-    }
-    public Room GetTop()
-    {
-        if(RoomController.instance.DoesRoomExist(X, Y + 1))
+        public Room GetTop()
         {
-            return RoomController.instance.FindRoom(X, Y + 1);
+            if(RoomController.instance.DoesRoomExist(X, Y + 1))
+            {
+                return RoomController.instance.FindRoom(X, Y + 1);
+            }
+            return null;
         }
-        return null;
-    }
-    public Room GetBottom()
-    {
-        if(RoomController.instance.DoesRoomExist(X, Y - 1))
+        public Room GetBottom()
         {
-            return RoomController.instance.FindRoom(X, Y - 1);
+            if(RoomController.instance.DoesRoomExist(X, Y - 1))
+            {
+                return RoomController.instance.FindRoom(X, Y - 1);
+            }
+            return null;
         }
-        return null;
-    }
 
 
-    void OnDrawGizmos()
-    {
-        Gizmos.color = Color.red;
-        Gizmos.DrawWireCube(transform.position, new Vector3(Width, Height, 0));
-    }
-
-    public Vector3 GetRoomCentre()
-    {
-        return new Vector3( X * Width, Y * Height);
-    }
-
-    void OnTriggerEnter2D(Collider2D other)
-    {
-        if(other.CompareTag("Player"))
+        void OnDrawGizmos()
         {
-            RoomController.instance.OnPlayerEnterRoom(this);
+            Gizmos.color = Color.red;
+            Gizmos.DrawWireCube(transform.position, new Vector3(Width, Height, 0));
+        }
+
+        public Vector3 GetRoomCentre()
+        {
+            return new Vector3( X * Width, Y * Height);
+        }
+
+        void OnTriggerEnter2D(Collider2D other)
+        {
+            if(other.CompareTag("Player"))
+            {
+                RoomController.instance.OnPlayerEnterRoom(this);
+            }
         }
     }
 }
