+++ Assets/Scripts/DungeonGeneration/RoomController.cs
@@ -1,235 +1,240 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
+using Game.Camera;
+using Game.Enemies;
 using UnityEngine;
 using UnityEngine.SceneManagement;
-using System.Linq;
 
-public class RoomInfo
+namespace DungeonGeneration
 {
-    public string name;
-    public int X;
-    public int Y;
-}
-
-public class RoomController : MonoBehaviour
-{
-
-    public static RoomController instance;
-
-    string currentWorldName = "Basement";
-
-    RoomInfo currentLoadRoomData;
-
-    Room currRoom;
-
-    Queue<RoomInfo> loadRoomQueue = new Queue<RoomInfo>();
-
-    public List<Room> loadedRooms = new List<Room>();
-
-    bool isLoadingRoom = false;
-    bool spawnedBossRoom = false;
-    bool updatedRooms = false;
-
-    void Awake()
+    public class RoomInfo
     {
-        instance = this;
+        public string name;
+        public int X;
+        public int Y;
     }
+
+    public class RoomController : MonoBehaviour
+    {
+
+        public static RoomController instance;
+
+        string currentWorldName = "Basement";
+
+        RoomInfo currentLoadRoomData;
+
+        Room currRoom;
+
+        Queue<RoomInfo> loadRoomQueue = new Queue<RoomInfo>();
+
+        public List<Room> loadedRooms = new List<Room>();
+
+        bool isLoadingRoom = false;
+        bool spawnedBossRoom = false;
+        bool updatedRooms = false;
+
+        void Awake()
+        {
+            instance = this;
+        }
     
-    void Update()
-    {
-        UpdateRoomQueue();
-    }
-
-    void UpdateRoomQueue()
-    {
-        if(isLoadingRoom)
+        void Update()
         {
-            return;
+            UpdateRoomQueue();
         }
 
-        if(loadRoomQueue.Count == 0)
+        void UpdateRoomQueue()
         {
-            if(!spawnedBossRoom)
+            if(isLoadingRoom)
             {
-                StartCoroutine(SpawnBossRoom());
-            } 
-            else if(spawnedBossRoom && !updatedRooms)
-            {
-                foreach(Room room in loadedRooms)
-                {
-                    room.RemoveUnconnectedDoors();
-                }
-                UpdateRooms();
-                updatedRooms = true;
-            }
-            return;
-        }
-
-        currentLoadRoomData = loadRoomQueue.Dequeue();
-        isLoadingRoom = true;
-
-        StartCoroutine(LoadRoomRoutine(currentLoadRoomData));
-    }
-
-    IEnumerator SpawnBossRoom()
-    {
-        spawnedBossRoom = true;
-        yield return new WaitForSeconds(0.5f);
-        if(loadRoomQueue.Count == 0)
-        {
-            Room bossRoom = loadedRooms[loadedRooms.Count - 1];
-            Room tempRoom = new Room(bossRoom.X, bossRoom.Y);
-            Destroy(bossRoom.gameObject);
-            var roomToRemove = loadedRooms.Single(r => r.X == tempRoom.X && r.Y == tempRoom.Y);
-            loadedRooms.Remove(roomToRemove);
-            LoadRoom("End", tempRoom.X, tempRoom.Y);
-        }
-    }
-
-    public void LoadRoom( string name, int x, int y)
-    {
-        if(DoesRoomExist(x, y))
-        {
-            return;
-        }
-
-        RoomInfo newRoomData = new RoomInfo();
-        newRoomData.name = name;
-        newRoomData.X = x;
-        newRoomData.Y = y;
-
-        loadRoomQueue.Enqueue(newRoomData);
-    }
-
-    IEnumerator LoadRoomRoutine(RoomInfo info)
-    {
-        string roomName = currentWorldName + info.name;
-
-        AsyncOperation loadRoom = SceneManager.LoadSceneAsync(roomName, LoadSceneMode.Additive);
-
-        while(loadRoom.isDone == false)
-        {
-            yield return null;
-        }
-    }
-
-    public void RegisterRoom(Room room)
-    {
-        if(!DoesRoomExist(currentLoadRoomData.X, currentLoadRoomData.Y))
-        {
-            room.transform.position = new Vector3(
-                currentLoadRoomData.X * room.Width,
-                currentLoadRoomData.Y * room.Height,
-                0
-            );
-
-            room.X = currentLoadRoomData.X;
-            room.Y = currentLoadRoomData.Y;
-            room.name = currentWorldName + "-" + currentLoadRoomData.name + " " + room.X + ", " + room.Y;
-            room.transform.parent = transform;
-
-            isLoadingRoom = false;
-
-            if (loadedRooms.Count == 0)
-            {
-                SetCamerasRoom(room);
+                return;
             }
 
-            loadedRooms.Add(room);
-        }
-        else
-        {
-            Destroy(room.gameObject);
-            isLoadingRoom = false;
-        }
-
-    }
-
-    public bool DoesRoomExist( int x, int y)
-    {
-        return loadedRooms.Find( item => item.X == x && item.Y == y) != null;
-    }
-
-    public Room FindRoom( int x, int y)
-    {
-        return loadedRooms.Find( item => item.X == x && item.Y == y);
-    }
-
-    public string GetRandomRoomName()
-    {
-        string[] possibleRooms = {
-            "Empty",
-            //"Basic1"
-        };
-
-        return possibleRooms[Random.Range(0, possibleRooms.Length)];
-    }
-
-    public void OnPlayerEnterRoom(Room room)
-    {
-        SetCamerasRoom(room);
-        currRoom = room;
-        StartCoroutine(RoomCoroutine());
-    }
-
-    public IEnumerator RoomCoroutine()
-    {
-        yield return new WaitForSeconds(0.2f);
-        UpdateRooms();
-    }
-
-    public void UpdateRooms()
-    {
-        foreach(Room room in loadedRooms)
-        {
-            //Not current rooms
-            if(currRoom != room)
+            if(loadRoomQueue.Count == 0)
             {
-                EnemyController[] enemies = room.GetComponentsInChildren<EnemyController>();
-                if(enemies != null)
+                if(!spawnedBossRoom)
                 {
-                    foreach(EnemyController enemy in enemies)
+                    StartCoroutine(SpawnBossRoom());
+                } 
+                else if(spawnedBossRoom && !updatedRooms)
+                {
+                    foreach(Room room in loadedRooms)
                     {
-                        enemy.notInRoom = true;
+                        room.RemoveUnconnectedDoors();
                     }
+                    UpdateRooms();
+                    updatedRooms = true;
                 }
-                foreach(Door door in room.GetComponentsInChildren<Door>())
-                {
-                    door.OpenDoor();
-                }
+                return;
             }
-            //Current room
+
+            currentLoadRoomData = loadRoomQueue.Dequeue();
+            isLoadingRoom = true;
+
+            StartCoroutine(LoadRoomRoutine(currentLoadRoomData));
+        }
+
+        IEnumerator SpawnBossRoom()
+        {
+            spawnedBossRoom = true;
+            yield return new WaitForSeconds(0.5f);
+            if(loadRoomQueue.Count == 0)
+            {
+                Room bossRoom = loadedRooms[loadedRooms.Count - 1];
+                Room tempRoom = new Room(bossRoom.X, bossRoom.Y);
+                Destroy(bossRoom.gameObject);
+                var roomToRemove = loadedRooms.Single(r => r.X == tempRoom.X && r.Y == tempRoom.Y);
+                loadedRooms.Remove(roomToRemove);
+                LoadRoom("End", tempRoom.X, tempRoom.Y);
+            }
+        }
+
+        public void LoadRoom( string name, int x, int y)
+        {
+            if(DoesRoomExist(x, y))
+            {
+                return;
+            }
+
+            RoomInfo newRoomData = new RoomInfo();
+            newRoomData.name = name;
+            newRoomData.X = x;
+            newRoomData.Y = y;
+
+            loadRoomQueue.Enqueue(newRoomData);
+        }
+
+        IEnumerator LoadRoomRoutine(RoomInfo info)
+        {
+            string roomName = currentWorldName + info.name;
+
+            AsyncOperation loadRoom = SceneManager.LoadSceneAsync(roomName, LoadSceneMode.Additive);
+
+            while(loadRoom.isDone == false)
+            {
+                yield return null;
+            }
+        }
+
+        public void RegisterRoom(Room room)
+        {
+            if(!DoesRoomExist(currentLoadRoomData.X, currentLoadRoomData.Y))
+            {
+                room.transform.position = new Vector3(
+                    currentLoadRoomData.X * room.Width,
+                    currentLoadRoomData.Y * room.Height,
+                    0
+                );
+
+                room.X = currentLoadRoomData.X;
+                room.Y = currentLoadRoomData.Y;
+                room.name = currentWorldName + "-" + currentLoadRoomData.name + " " + room.X + ", " + room.Y;
+                room.transform.parent = transform;
+
+                isLoadingRoom = false;
+
+                if (loadedRooms.Count == 0)
+                {
+                    SetCamerasRoom(room);
+                }
+
+                loadedRooms.Add(room);
+            }
             else
             {
-              EnemyController[] enemies = room.GetComponentsInChildren<EnemyController>();
-                if(enemies.Length > 0)
+                Destroy(room.gameObject);
+                isLoadingRoom = false;
+            }
+
+        }
+
+        public bool DoesRoomExist( int x, int y)
+        {
+            return loadedRooms.Find( item => item.X == x && item.Y == y) != null;
+        }
+
+        public Room FindRoom( int x, int y)
+        {
+            return loadedRooms.Find( item => item.X == x && item.Y == y);
+        }
+
+        public string GetRandomRoomName()
+        {
+            string[] possibleRooms = {
+                "Empty",
+                //"Basic1"
+            };
+
+            return possibleRooms[Random.Range(0, possibleRooms.Length)];
+        }
+
+        public void OnPlayerEnterRoom(Room room)
+        {
+            SetCamerasRoom(room);
+            currRoom = room;
+            StartCoroutine(RoomCoroutine());
+        }
+
+        public IEnumerator RoomCoroutine()
+        {
+            yield return new WaitForSeconds(0.2f);
+            UpdateRooms();
+        }
+
+        public void UpdateRooms()
+        {
+            foreach(Room room in loadedRooms)
+            {
+                //Not current rooms
+                if(currRoom != room)
                 {
-                    foreach(EnemyController enemy in enemies)
+                    EnemyController[] enemies = room.GetComponentsInChildren<EnemyController>();
+                    if(enemies != null)
                     {
-                        enemy.notInRoom = false;
+                        foreach(EnemyController enemy in enemies)
+                        {
+                            enemy.notInRoom = true;
+                        }
                     }
-                    
-                    foreach(Door door in room.GetComponentsInChildren<Door>())
-                    {
-                        door.CloseDoor();
-                    }
-                }
-                else
-                {
                     foreach(Door door in room.GetComponentsInChildren<Door>())
                     {
                         door.OpenDoor();
                     }
-                }  
+                }
+                //Current room
+                else
+                {
+                    EnemyController[] enemies = room.GetComponentsInChildren<EnemyController>();
+                    if(enemies.Length > 0)
+                    {
+                        foreach(EnemyController enemy in enemies)
+                        {
+                            enemy.notInRoom = false;
+                        }
+                    
+                        foreach(Door door in room.GetComponentsInChildren<Door>())
+                        {
+                            door.CloseDoor();
+                        }
+                    }
+                    else
+                    {
+                        foreach(Door door in room.GetComponentsInChildren<Door>())
+                        {
+                            door.OpenDoor();
+                        }
+                    }  
+                }
             }
         }
-    }
 
-    void SetCamerasRoom(Room room)
-    {
-        foreach (Camera cam in Camera.allCameras)
+        void SetCamerasRoom(Room room)
         {
-            cam.GetComponent<CameraController>().currRoom = room;
+            foreach (Camera cam in Camera.allCameras)
+            {
+                cam.GetComponent<CameraController>().currRoom = room;
+            }
         }
     }
 }
