--- Assets/Scripts/Player/CollectionController.cs
@@ -1,40 +1,0 @@
-﻿using UnityEngine;
-
-[System.Serializable]
-public class Item
-{
-    public string name;
-    public string description;
-    public Sprite itemImage;
-}
-
-public class CollectionController : MonoBehaviour
-{
-
-    public Item item;
-    public float healthChange;
-    public float moveSpeedChange;
-    public float attackSpeedChange;
-    public float bulletSizeChange;
-    // Start is called before the first frame update
-    void Start()
-    {
-        GetComponent<SpriteRenderer>().sprite = item.itemImage;
-        Destroy(GetComponent<PolygonCollider2D>());
-        gameObject.AddComponent<PolygonCollider2D>();
-    }
-
-    private void OnTriggerEnter2D(Collider2D collision)
-    {
-        if(collision.tag == "Player")
-        {
-            PlayerController.collectedAmount++;
-            GameController.HealPlayer(healthChange);
-            GameController.MoveSpeedChange(moveSpeedChange);
-            GameController.FireRateChange(attackSpeedChange);
-            GameController.BulletSizeChange(bulletSizeChange);
-            GameController.instance.UpdateCollectedItems(this);
-            Destroy(gameObject);
-        }
-    }
-}
