+++ Assets/Scripts/ScriptableObjects/FamiliarData.cs
@@ -1,10 +1,13 @@
 ﻿using UnityEngine;
 
-[CreateAssetMenu(fileName = "Familiar.asset", menuName = "Familiars/FamiliarObject")]
-public class FamiliarData : ScriptableObject
+namespace ScriptableObjects
 {
-    public string familirType;
-    public float speed;
-    public float fireDelay;
-    public GameObject bulletPrefab;
+    [CreateAssetMenu(fileName = "Familiar.asset", menuName = "Familiars/FamiliarObject")]
+    public class FamiliarData : ScriptableObject
+    {
+        public string familirType;
+        public float speed;
+        public float fireDelay;
+        public GameObject bulletPrefab;
+    }
 }
