+++ Assets/Scripts/ScriptableObjects/Spawners/SpawnerData.cs
@@ -1,9 +1,12 @@
 ﻿using UnityEngine;
 
-[CreateAssetMenu(fileName = "Spawner.asset", menuName = "Spawners/Spawner")]
-public class SpawnerData : ScriptableObject
+namespace ScriptableObjects.Spawners
 {
-    public GameObject itemToSpawn;
-    public int minSpawn;
-    public int maxSpawn;
+    [CreateAssetMenu(fileName = "Spawner.asset", menuName = "Spawners/Spawner")]
+    public class SpawnerData : ScriptableObject
+    {
+        public GameObject itemToSpawn;
+        public int minSpawn;
+        public int maxSpawn;
+    }
 }
